requirements.txt
router
requests
ollama
python-dotenv
langchain_ollama
langchain_core